Extract int/float toString
