+++ runtime/arithmetic/CMakeLists.txt
@@ -3,6 +3,10 @@ add_library(arithmetic STATIC
   float.cpp
 )
 
+target_link_libraries(arithmetic
+  numeric_strings
+)
+
 install(
   TARGETS arithmetic
   ARCHIVE DESTINATION lib/kllvm
